test: HintGenerator의 예외처리 테스트들 PitcherTest로 옮기기
